+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+23.10.1
+-------
+
+* Fixed issue with refreshable credentials not working due to timing issue introduced in 23.10.
+* Improved performance of :py:class:`aiodynamo.credentials.ChainCredentials`
+
 23.10
 -----